requirements.txt
fastapi==0.95.1
jose==1.0.0
passlib==1.7.4
pydantic==1.10.7
python_jose==3.3.0
SQLAlchemy==2.0.9
pymysql==1.0.3
uvicorn==0.22.0
bcrypt==4.0.1
numpy==1.21.6
pillow==9.5.0
opencv-python==4.7.0.72